+++ tests/Support/Model/CastPayload.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace UIAwesome\Model\Tests\Support\Model;
 
-use UIAwesome\Model\AbstractModel;
 use UIAwesome\Model\Attribute\{Cast, MapFrom, Trim};
+use UIAwesome\Model\BaseModel;
 use UIAwesome\Model\Tests\Support\Caster\PipeSeparatedCaster;
 
 /**
@@ -14,7 +14,7 @@ use UIAwesome\Model\Tests\Support\Caster\PipeSeparatedCaster;
  * @copyright Copyright (C) 2026 Terabytesoftw.
  * @license https://opensource.org/license/bsd-3-clause BSD 3-Clause License.
  */
-final class CastPayload extends AbstractModel
+final class CastPayload extends BaseModel
 {
     #[Cast(PipeSeparatedCaster::class)]
     public array $keywords = [];