+++ tests/Support/Model/DefaultValueParent.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace UIAwesome\Model\Tests\Support\Model;
 
-use UIAwesome\Model\AbstractModel;
 use UIAwesome\Model\Attribute\{DefaultValue, DoNotCollect};
+use UIAwesome\Model\BaseModel;
 
 /**
  * Stub parent model combining `DoNotCollect` and `DefaultValue` metadata for tests.
@@ -13,7 +13,7 @@ use UIAwesome\Model\Attribute\{DefaultValue, DoNotCollect};
  * @copyright Copyright (C) 2026 Terabytesoftw.
  * @license https://opensource.org/license/bsd-3-clause BSD 3-Clause License.
  */
-class DefaultValueParent extends AbstractModel
+class DefaultValueParent extends BaseModel
 {
     #[DoNotCollect]
     #[DefaultValue('parent-default')]