+++ tests/Support/Model/NonNamespaced.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-use UIAwesome\Model\AbstractModel;
+use UIAwesome\Model\BaseModel;
 
 /**
  * Stub non-namespaced model used for tests.
@@ -10,4 +10,4 @@ use UIAwesome\Model\AbstractModel;
  * @copyright Copyright (C) 2024 Terabytesoftw.
  * @license https://opensource.org/license/bsd-3-clause BSD 3-Clause License.
  */
-final class NonNamespaced extends AbstractModel {}
+final class NonNamespaced extends BaseModel {}