+++ tests/Support/Model/UnionType.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace UIAwesome\Model\Tests\Support\Model;
 
-use UIAwesome\Model\AbstractModel;
+use UIAwesome\Model\BaseModel;
 
 /**
  * Stub model with union-typed property used for tests.
@@ -12,7 +12,7 @@ use UIAwesome\Model\AbstractModel;
  * @copyright Copyright (C) 2024 Terabytesoftw.
  * @license https://opensource.org/license/bsd-3-clause BSD 3-Clause License.
  */
-final class UnionType extends AbstractModel
+final class UnionType extends BaseModel
 {
     private bool|int|object|string|null $union = null;
 }